fix(artist)!: Change type of popularity to double?

+++ lib/src/api_models/artists/response.dart
@@ -16,7 +16,7 @@ class Artist {
   ///
   /// The artist's popularity is calculated from the popularity of all the
   /// artist's tracks.
-  final int? popularity;
+  final double? popularity;
 
   Artist({
     required this.id,
